+++ frontend/react/src/styles/App.css
